moved and renamed sqe package to original location for rmi
