+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>at.fhhagenberg.sqe</groupId>
+    <groupId>sqe</groupId>
     <artifactId>elevator-control-center-team-d</artifactId>
     <version>0.0.1-SNAPSHOT</version>
 
@@ -9,8 +9,8 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>11</maven.compiler.source>
         <maven.compiler.target>11</maven.compiler.target>
-        <sonar.projectKey>fhhagenberg-sqe-mcm-ws20_elevator-control-center-team-d</sonar.projectKey>
-        <sonar.organization>fhhagenberg-sqe-mcm-ws20</sonar.organization>
+        <sonar.projectKey>fhhagenberg-sqelevator-mcm-ws20_elevator-control-center-team-d</sonar.projectKey>
+        <sonar.organization>fhhagenberg-sqelevator-mcm-ws20</sonar.organization>
         <sonar.host.url>https://sonarcloud.io</sonar.host.url>
     </properties>
