+++ src/test/java/sqelevator/GUIApplicationTest.java
@@ -1,4 +1,4 @@
-package at.fhhagenberg.sqe;
+package sqelevator;
 
 import at.fhhagenberg.elevatorsys.ControlCenter;
 import at.fhhagenberg.elevatorsys.view.ControlCenterUI;
